requirements.txt
numpy
pillow
matplotlib
pygame
keyboard
requests
scipy
pywin32
opencv-python